Update 챗봇.js
json 활용, 말 가르치기 기능 추가

+++ 챗봇.js
@@ -1,6 +1,14 @@
 var n = 0;
 var follow = 0;
 
+var image = document.getElementById("image");
+var body = document.getElementById("body");
+var saysBox = document.getElementById("saysBox");
+
+var question = "";
+var answer = "";
+var key = 0;
+
 var app = document.getElementById('hamConsole');
 
 var typewriter = new Typewriter(app, {
@@ -11,13 +19,28 @@ typewriter.typeString("쮝슨과 대화를 해보세요!")
     .pauseFor(1900)
     .start()
 
-function checkText() {
-    var hamSays = document.getElementsByClassName("hamSays");
-    var image = document.getElementById("image");
-    var body = document.getElementById("body");
-    var saysBox = document.getElementById("saysBox");
+var json = [
+    {
+        "question":"지금 뭐해?",
+        "answer":"팬케이크를 먹고 있습니다..!!"
+    },
+    {
+        "question":"쮝슨이 제일 좋아하는 거는?",
+        "answer":"먹을 걸 제일 좋아합니다!"
+    },
+    {
+        "question":"너의 매력포인트는 뭐야?",
+        "answer":"입술이라고 생각합니다..!!"
+    },
+    {
+        "question":"이제 가볼게",
+        "answer":"다음에도 저랑 대화해 주세요!!"
+    }
+]
+
+function inputText() {
     var value = document.getElementById("console").value;
-    console.log(value);
+
 
     if (follow == 1) {
         if (value.includes("그만")) {
@@ -39,6 +62,7 @@ function checkText() {
     }
 
     else if (follow == 0) {
+
         if (value.includes("따라")) {
             typewriter
                 .deleteAll()
@@ -73,15 +97,6 @@ function checkText() {
     
         else if (value=="") {
             alert("쮝슨에게 말을 걸어주세요!");
-        }
-    
-        else if (value=="지금 뭐해?") {
-            typewriter
-                .deleteAll()
-                .typeString("팬케이크를 먹고 있습니다..!!")
-                .pauseFor(1900)
-                .start()
-            image.src = "./static/images/쮝슨_맛있다.png"
         }
     
         else if (value.includes("불") && value.includes("꺼")) {
@@ -148,24 +163,91 @@ function checkText() {
     
         }
     
-        else if (value=="이제 가볼게") {
+        else {
+            checkText();
+        }
+    }
+}
+
+function checkText() {
+    var value = document.getElementById("console").value;
+
+    if (key == 1) {
+        if (value == "응") {
             typewriter
                 .deleteAll()
-                .typeString("다음에도 저랑 대화해 주세요!!")
+                .typeString("대답을 입력해 주세요.")
                 .pauseFor(1900)
                 .start()
-            image.src = "./static/images/쮝슨.png";
-            setTimeout(reload, 5800);
+            key = 2;
         }
-    
         else {
             typewriter
                 .deleteAll()
-                .typeString("무슨 뜻인지 잘 모르겠습니다,,,")
+                .typeString("쮝!")
                 .pauseFor(1900)
                 .start()
+            key = 0;
+        }
+        return;
+    }
+
+    if (key == 2) {
+        answer = value;
+        push_json();
+        return;
+    }
+
+    for (let i=0;i<json.length;i++) {
+        if (value == json[0].question) {
+            image.src = "./static/images/쮝슨_맛있다.png"
+            typewriter
+                .deleteAll()
+                .typeString(json[0].answer)
+                .pauseFor(1900)
+                .start()
+            return;
+        }
+
+        else if (value == json[3].question) {
+            image.src = "./static/images/쮝슨.png";
+            typewriter
+                .deleteAll()
+                .typeString(json[3].answer)
+                .pauseFor(1900)
+                .start()
+            setTimeout(reload, 5800);
+            return;
+        }
+
+        else if (value == json[i].question) {
+            image.src = "./static/images/쮝슨.png";
+            typewriter
+                .deleteAll()
+                .typeString(json[i].answer)
+                .pauseFor(1900)
+                .start()
+            return;
         }
     }
+
+    typewriter
+        .deleteAll()
+        .typeString("말을 가르쳐 주실래요?(응 or 아니)")
+        .pauseFor(1900)
+        .start()
+    question = value;
+    key = 1;
+}
+
+function push_json() {
+    json.push({question:`${question}`, answer:`${answer}`});
+    typewriter
+        .deleteAll()
+        .typeString("말을 배웠습니다!!")
+        .pauseFor(1900)
+        .start()
+    key = 0;
 }
 
 function reload() {
@@ -173,7 +255,7 @@ function reload() {
 }
 
 function onClick1() {
-    checkText();
+    inputText();
 }
 
 document.getElementById("send").addEventListener("click", onClick1)
